+++ src/main/java/eu/unicore/uftp/standalone/authclient/AuthRequest.java
@@ -7,7 +7,8 @@ public class AuthRequest {
 	public String serverPath;
 	public int streamCount = 1;
 	public boolean compress = false;
-	public String encryptionKey = null;  
+	public String encryptionKey = null;
+	public String encryptionAlgorithm = null;
 	public String group = null;
 	public String client = null;
 	public boolean persistent = true;